+++ src/test/java/domain/DealerTest.java
@@ -17,4 +17,17 @@ public class DealerTest {
         assertThat(dealer.isBelowThreshold()).isFalse();
     }
 
+    @Test
+    @DisplayName("딜러의 핸드 총합이 16 이하면 카드를 추가로 받는다.")
+    void test2() {
+        Dealer dealer = new Dealer();
+
+        dealer.addCard(new Card(Denomination.TWO, Suit.CLUB));
+        dealer.addCard(new Card(Denomination.TEN, Suit.DIAMOND));
+
+        dealer.drawWithThreshold(new Deck());
+
+        assertThat(dealer.getCardCount()).isGreaterThan(2);
+    }
+
 }
